Avoid calling eglMakeCurrent prior to creating a window surface.
If we call eglMakeCurrent with no draw or read surface prior to creating the window surface, then on Nvidia/X11 we get a black window. This change avoids making the context current to read out the EGLConfig used to create the window surface, and instead extracts it directly from the raw EGLContext.

Fixes #173