Fix clippy lint `mismatched_target_os`
Fix the deny-by-default clippy lint `mismatched_target_os`.
We rename the `linux` cfg option to `free_unix`, which
more accurately represents how it is used. The naming follows the
cfg names chosen by the `winit` crate.
the `wayland` cfg is slightly adjusted, to only be selected as
the default if 1. wayland is enabled AND 2. either x11 is disabled,
or the `sm-wayland-default` is selected

Example clippy error message:
```
error: operating system used in target family position
  --> surfman/src/platform/unix/mod.rs:13:1
   |
13 | #[cfg(linux)]
   | ^^^^^^-----^^
   |       |
   |       help: try: `target_os = "linux"`
   |
   = help: did you mean `unix`?
   = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#mismatched_target_os
 ```

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>