Use cstr literals where possible
Convert usage of byte str and normal str with a `\0`
to cstr literals where possible.
Since CStr::as_ptr() returns `*const c_char` this also allows
removing a nice amount of pointer casts.

On Windows this also allows removing casts to [`LPCSTR`], which
is defined as a `*const CHAR`, where `CHAR` is defined as `c_char`
and `c_char` is taken from the std library if the `std` feature is
enabled.

[`LPCSTR`]: https://docs.rs/winapi/latest/winapi/um/winnt/type.LPCSTR.html

Signed-off-by: Jonathan Schwender <jonathan.schwender@huawei.com>